pyproject.toml
[tool.poetry]
name = "imiit-student-bot"
version = "1.0.0"
description = "Imiit Student Bot"
authors = ["Николай Иванов <nikolai.in@tuta.io>"]
license = "GPL-3.0"
readme = "README.rst"
homepage = "https://github.com/nikolai-in/imiit-student-bot"
repository = "https://github.com/nikolai-in/imiit-student-bot"
documentation = "https://imiit-student-bot.readthedocs.io"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/nikolai-in/imiit-student-bot/releases"

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0.1"
python-telegram-bot = "^13.8.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = {extras = ["toml"], version = "^6.1"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.13.0"
xdoctest = {extras = ["colors"], version = "^0.15.10"}
sphinx = "^4.2.0"
sphinx-autobuild = "^2021.3.14"
pre-commit = "^2.15.0"
flake8 = "^4.0.1"
black = "^21.10b0"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.9.2"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.1"
darglint = "^1.8.1"
reorder-python-imports = "^2.6.0"
pre-commit-hooks = "^4.0.1"
sphinx-rtd-theme = "^1.0.0"
sphinx-click = "^3.0.2"
Pygments = "^2.10.0"
furo = "^2021.11.23"

[tool.poetry.scripts]
imiit-student-bot = "imiit_student_bot.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["imiit_student_bot"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
